# backend/src/docs/swagger.yaml
basePath: /v1
definitions:
...
  model.Address:
    properties:
      address:
        type: string
      apartment:
        type: string
      country:
        type: string
      createdAt:
        type: string
      district:
        type: string
      fullName:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      phone:
        type: string
      province:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
      ward:
        type: string
    type: object
  model.CartItem:
    properties:
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      quantity:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.Category:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  model.Image:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isPrimary:
        type: boolean
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      url:
        type: string
    type: object
  model.Order:
    properties:
      createdAt:
        type: string
      id:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      paymentMethod:
        $ref: '#/definitions/model.PaymentMethod'
      shippingAddress:
        type: string
      status:
        $ref: '#/definitions/model.OrderStatus'
      total:
        type: number
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.OrderItem:
    properties:
      id:
        type: string
      order:
        $ref: '#/definitions/model.Order'
      orderId:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      quantity:
        type: integer
    type: object
  model.OrderStatus:
    enum:
    - PENDING
    - PAID
    - SHIPPED
    - DELIVERED
    - CANCELLED
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusPaid
    - OrderStatusShipped
    - OrderStatusDelivered
    - OrderStatusCancelled
  model.PaymentMethod:
    enum:
    - COD
    - BANK_TRANSFER
    - STRIPE
    - MOMO
    - VNPAY
    type: string
    x-enum-varnames:
    - PaymentMethodCOD
    - PaymentMethodBankTransfer
    - PaymentMethodStripe
    - PaymentMethodMomo
    - PaymentMethodVnpay
  model.Product:
    properties:
      cartItems:
        items:
          $ref: '#/definitions/model.CartItem'
        type: array
      category:
        $ref: '#/definitions/model.Category'
      categoryId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      discount:
        type: number
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      name:
        type: string
      orderItems:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      price:
        type: number
      quantity:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
      wishlist:
        items:
          $ref: '#/definitions/model.Wishlist'
        type: array
    type: object
  model.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      rating:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
  model.User:
    properties:
      address:
        type: string
      addresses:
        items:
          $ref: '#/definitions/model.Address'
        type: array
      bio:
        type: string
      cartItems:
        items:
          $ref: '#/definitions/model.CartItem'
        type: array
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isAdmin:
        type: boolean
      lastLoginAt:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
      phone:
        type: string
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      updatedAt:
        type: string
      wishlist:
        items:
          $ref: '#/definitions/model.Wishlist'
        type: array
    type: object
  model.Wishlist:
    properties:
      createdAt:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      productId:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@birdnest.vn
    name: API Support
  description: BirdNest Shop REST API for e-commerce platform
  license:
    name: MIT
    url: https://github.com/indrayyana/go-fiber-boilerplate/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: BirdNest Shop API
  version: 1.0.0
paths:
...
  /admin/orders/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an order (admin only)
      tags:
      - Orders
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update order status (admin only)
      tags:
      - Orders
  /admin/orders/stats:
    get:
      consumes:
      - application/json
      parameters:
      - default: daily
        description: Statistics period (daily, weekly, monthly, yearly)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order statistics (admin only)
      tags:
      - Orders
...
  /orders:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Order status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Orders with pagination
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user's orders or all orders (admin)
      tags:
      - Orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order data with items, shipping address, and payment method
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific order
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"